Make `closed` flag of the resource static
Issue: #2816
